# contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Shadowcook API
  version: 0.1.0
paths:
...
  /cookbook:
    get:
      operationId: getCookbook
      summary: Returns published public recipes and, for an authenticated user, recipes in the user's tenant memberships.
      responses:
        '200':
          description: The accessible cookbook overview.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [tenant, categories, recipes, canManageCategories]
                properties:
                  tenant:
                    oneOf:
                      - type: 'null'
                      - type: object
                        additionalProperties: false
                        required: [display_name]
                        properties:
                          display_name: { type: string }
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/CookbookCategory'
                  recipes:
                    type: array
                    items:
                      $ref: '#/components/schemas/CookbookRecipe'
                  canManageCategories:
                    type: boolean
        '403':
          description: A password change is required before accessing the cookbook.
          content:
            application/json:
              schema:
              $ref: '#/components/schemas/ApiError'
        '404':
          description: The requested tenant is disabled or does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /cookbook/recipes/{publicId}:
    get:
      operationId: getCookbookRecipe
      summary: Returns one accessible published recipe with its step-specific ingredient usages.
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The accessible published recipe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CookbookRecipeDetail'
        '403':
          description: A password change is required before accessing the recipe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The recipe is not accessible or does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
...
  schemas:
...
    CookbookCategory:
      type: object
      additionalProperties: false
      required: [public_id, parent_public_id, name, slug, sort_order]
      properties:
        public_id:
          type: string
          format: uuid
        parent_public_id:
          type: [string, 'null']
          format: uuid
        name:
          type: string
        slug:
          type: string
        sort_order:
          type: integer
          minimum: 0
    CookbookRecipe:
      type: object
      additionalProperties: false
      required: [public_id, slug, title, summary, category_public_ids]
      properties:
        public_id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        summary:
          type: [string, 'null']
        category_public_ids:
          type: array
          items:
            type: string
            format: uuid
    CookbookRecipeDetail:
      type: object
      additionalProperties: false
      required: [public_id, slug, title, summary, steps]
      properties:
        public_id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        summary:
          type: [string, 'null']
        steps:
          type: array
          items:
            $ref: '#/components/schemas/CookbookRecipeStep'
    CookbookRecipeStep:
      type: object
      additionalProperties: false
      required: [public_id, sort_order, instruction, ingredients]
      properties:
        public_id:
          type: string
          format: uuid
        sort_order:
          type: integer
          minimum: 0
        instruction:
          type: string
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/CookbookIngredientUsage'
    CookbookIngredientUsage:
      type: object
      additionalProperties: false
      required:
        [
          sort_order,
          amount,
          unit_symbol,
          unit_localization_key,
          ingredient_name,
          ingredient_localization_key,
          special_kind,
          note,
          is_optional,
        ]
      properties:
        sort_order:
          type: integer
          minimum: 0
        amount:
          type: [string, 'null']
        unit_symbol:
          type: [string, 'null']
        unit_localization_key:
          type: [string, 'null']
        ingredient_name:
          type: string
        ingredient_localization_key:
          type: [string, 'null']
        special_kind:
          type: [string, 'null']
          enum: [NO_ICON, REMOVE, ADD, INFO, IMPORTANT, COOK, COOL, HEAT, WAIT, WORK_STEP, null]
        note:
          type: [string, 'null']
        is_optional:
          type: boolean
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, error]
      properties:
        code:
          type: string
          enum:
            - AUTHENTICATION_FAILED
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - INVALID_LOGIN_BODY
            - INVALID_PASSWORD
            - PASSWORD_CHANGE_REQUIRED
            - RECIPE_NOT_FOUND
            - RECIPE_CONFLICT
            - INVALID_RECIPE
            - INVALID_RECIPE_CATEGORY
            - PUBLISH_CATEGORY_REQUIRED
            - RECIPE_NOT_PUBLISHED
            - INVALID_UNIT
            - UNIT_CONFLICT
            - UNIT_NOT_FOUND
            - UNIT_IN_USE
            - INVALID_CONVERSION
            - UNIT_DIMENSION_MISMATCH
            - INVALID_INGREDIENT
            - INGREDIENT_CONFLICT
            - INGREDIENT_NOT_FOUND
            - INGREDIENT_IN_USE
            - INVALID_ALIAS
            - ALIAS_CONFLICT
            - ALIAS_NOT_FOUND
        error:
          type: string
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shadowcook_session